requirements.txt
Appium-Python-Client==2.11.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.1.0
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
pluggy==1.2.0
protobuf==4.21.12
pycparser==2.21
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
